Combine all themes into a single playground (#1539)
* move playground to new folder

* move playground to packages and standardize playground structure

* don't rely on bootstrap in playground

* fix styling, move share button to top right

* remove footer things

* more fixes

* encode theme in playground hash

* fix: don't switch to default theme when changing tabs

* correct package name is @rjsf/playground

* fix prod config for playground

* fix more configs

* use webpack-dev-server

* render form in a Frame

* add dependencies

* use DemoFrame component to properly fix material-ui styles

* use old bootstrap look and feel for the playground itself

* fix dist

* move netlify toml

* update netlify

* empty commit

* remove npm prepare to speed up travis runs

* fix webpack dist config for playground

* fix import

* disable minicssextractplugin to try to fix netlify

* fix netlify path

* no need to build playground / gh-pages for core package

* add HtmlWebpackPlugin
