+++ netlify.toml
@@ -1,0 +1,7 @@
+[build]
+  base = "packages/playground"
+  publish = "build"
+  command = "npm run build:playground"
+
+[build.environment]
+  SHOW_NETLIFY_BADGE = "true"
