+++ packages/playground/.eslintrc
@@ -1,0 +1,38 @@
+{
+    "parser": "babel-eslint",
+    "rules": {
+        "react/jsx-uses-react": 2,
+        "react/jsx-uses-vars": 2,
+        "react/react-in-jsx-scope": 2,
+        "react/jsx-tag-spacing": [1, {
+          "beforeSelfClosing": "always"
+        }],
+        "curly": [2],
+        "linebreak-style": [2, "unix"],
+        "semi": [2, "always"],
+        "comma-dangle": [0],
+        "no-unused-vars": [2, {
+          "vars": "all",
+          "args": "none",
+          "ignoreRestSiblings": true
+        }],
+        "no-console": [0],
+        "object-curly-spacing": [2, "always"],
+        "keyword-spacing": ["error"]
+    },
+    "env": {
+        "es6": true,
+        "browser": true,
+        "node": true
+    },
+    "extends": "eslint:recommended",
+    "ecmaFeatures": {
+        "modules": true,
+        "jsx": true,
+        "experimentalObjectRestSpread": true
+    },
+    "plugins": [
+        "jsx-a11y",
+        "react"
+    ]
+}
