+++ packages/playground/package.json
@@ -1,22 +1,20 @@
 {
-  "name": "@rjsf/core",
-  "version": "2.0.0-alpha.2",
-  "description": "A simple React component capable of building HTML forms out of a JSON schema.",
+  "name": "@rjsf/playground",
+  "version": "2.0.0-alpha.0",
+  "description": "rjsf playground",
   "scripts": {
     "build:lib": "rimraf lib && cross-env NODE_ENV=production babel -d lib/ src/",
     "build:dist": "rimraf dist && cross-env NODE_ENV=production webpack --config webpack.config.dist.js",
+    "build:playground": "rimraf build && cross-env NODE_ENV=production webpack --config webpack.config.prod.js",
     "cs-check": "prettier -l \"{playground,src,test}/**/*.js\"",
     "cs-format": "prettier \"{playground,src,test}/**/*.js\" --write",
     "dist": "npm run build:lib && npm run build:dist",
     "lint": "eslint src test playground",
     "precommit": "lint-staged",
+    "publish-to-gh-pages": "npm run build:playground && gh-pages --dist build/",
     "publish-to-npm": "npm run dist && npm publish",
     "preversion": "npm run build:playground && npm run dist && npm run cs-check && npm run lint",
-    "start": "node devServer.js",
-    "tdd": "cross-env NODE_ENV=test mocha --require @babel/register --watch --require ./test/setup-jsdom.js test/**/*_test.js",
-    "test": "cross-env NODE_ENV=test mocha --require @babel/register --require ./test/setup-jsdom.js test/**/*_test.js",
-    "test-coverage": "cross-env NODE_ENV=test nyc --reporter=lcov mocha --require @babel/register --require ./test/setup-jsdom.js test/**/*_test.js",
-    "test-debug": "cross-env NODE_ENV=test mocha --require @babel/register --require ./test/setup-jsdom.js --debug-brk --inspect test/Form_test.js"
+    "start": "webpack-dev-server --open --config webpack.config.dev.js"
   },
   "lint-staged": {
     "{playground,src,test}/**/*.js": [
@@ -43,10 +41,15 @@
     "ajv": "^6.7.0",
     "core-js": "^2.5.7",
     "json-schema-merge-allof": "^0.6.0",
+    "jss": "^10.0.3",
     "lodash": "^4.17.15",
     "prop-types": "^15.7.2",
     "react-app-polyfill": "^1.0.4",
+    "react-bootstrap": "^0.33.1",
+    "react-frame-component": "^4.1.1",
     "react-is": "^16.9.0",
+    "react-jsonschema-form": "^2.0.0-alpha.1",
+    "rjsf-material-ui": "^0.3.9",
     "shortid": "^2.2.14"
   },
   "devDependencies": {
@@ -54,11 +57,14 @@
     "@babel/core": "^7.4.5",
     "@babel/plugin-proposal-class-properties": "^7.4.4",
     "@babel/plugin-proposal-object-rest-spread": "^7.4.4",
+    "@babel/plugin-syntax-dynamic-import": "^7.7.4",
     "@babel/plugin-transform-react-jsx": "^7.3.0",
     "@babel/plugin-transform-runtime": "^7.4.4",
     "@babel/preset-env": "^7.4.5",
     "@babel/preset-react": "^7.0.0",
     "@babel/register": "^7.4.4",
+    "@material-ui/core": "^4.8.3",
+    "@material-ui/icons": "^4.5.1",
     "atob": "^2.0.3",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^8.0.6",
@@ -78,6 +84,7 @@
     "express": "^4.14.0",
     "gh-pages": "^0.11.0",
     "html": "^1.0.0",
+    "html-webpack-plugin": "^3.2.0",
     "husky": "^0.13.2",
     "jsdom": "^8.3.0",
     "json-loader": "^0.5.7",
@@ -97,11 +104,14 @@
     "react-transform-hmr": "^1.0.1",
     "redbox-react": "^1.3.3",
     "rimraf": "^2.5.4",
+    "sass": "^1.23.7",
+    "sass-loader": "^7.3.1",
     "sinon": "^7.5.0",
     "style-loader": "^0.13.1",
     "webpack": "^4.20.2",
     "webpack-cli": "^3.1.2",
     "webpack-dev-middleware": "^3.4.0",
+    "webpack-dev-server": "^3.10.1",
     "webpack-hot-middleware": "^2.13.2"
   },
   "directories": {
