+++ packages/playground/playground/index.prod.html
@@ -6,9 +6,10 @@
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <title>react-jsonschema-form playground</title>
   <link rel="stylesheet" id="theme" href="">
+  <link rel="stylesheet" href="./styles.css">
 </head>
 <body>
   <div id="app"></div>
-  <script src="/static/bundle.js"></script>
+  <script src="./bundle.js"></script>
 </body>
 </html>
