+++ packages/playground/src/samples/additionalProperties.js
