+++ packages/playground/src/samples/allOf.js
