+++ packages/playground/src/samples/alternatives.js
