+++ packages/playground/src/samples/anyOf.js
