+++ packages/playground/src/samples/arrays.js
