+++ packages/playground/src/samples/customArray.js
