+++ packages/playground/src/samples/customObject.js
