+++ packages/playground/src/samples/errors.js
