+++ packages/playground/src/samples/large.js
