+++ packages/playground/src/samples/numbers.js
