+++ packages/playground/src/samples/propertyDependencies.js
