+++ packages/playground/src/samples/schemaDependencies.js
