+++ packages/playground/src/samples/simple.js
