+++ packages/playground/src/samples/single.js
