+++ packages/playground/webpack.config.dev.js
@@ -1,0 +1,59 @@
+var path = require("path");
+var webpack = require("webpack");
+const MonacoWebpackPlugin = require('monaco-editor-webpack-plugin');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+module.exports = {
+  mode: "development",
+  devtool: "source-map",
+  entry: [
+    "webpack-hot-middleware/client?reload=true",
+    "./playground/app"
+  ],
+  output: {
+    path: path.join(__dirname, "build"),
+  },
+  devServer: {
+    contentBase: path.join(__dirname, 'build'),
+    compress: true,
+    port: 8080
+  },
+  plugins: [
+    new MonacoWebpackPlugin({
+      languages: ['json']
+    }),
+    new webpack.HotModuleReplacementPlugin(),
+    new HtmlWebpackPlugin({
+      filename: 'index.html',
+      template: 'playground/index.html'
+    }),
+  ],
+  module: {
+    rules: [
+      {
+        test: /\.jsx?$/,
+        use: [
+          "babel-loader",
+        ],
+        include: [
+          path.join(__dirname, "src"),
+          path.join(__dirname, "playground"),
+          path.join(__dirname, "node_modules", "codemirror", "mode", "javascript"),
+        ]
+      },
+      {
+        test: /\.s?css$/,
+        use: [
+          "style-loader",
+          "css-loader",
+          "sass-loader"
+        ],
+        include: [
+          path.join(__dirname, "src"),
+          path.join(__dirname, "playground"),
+          path.join(__dirname, "node_modules"),
+        ],
+      },
+    ]
+  }
+};
