+++ packages/playground/webpack.config.dist.js
@@ -1,31 +1,26 @@
 var path = require("path");
 var webpack = require("webpack");
-var MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const MonacoWebpackPlugin = require('monaco-editor-webpack-plugin');
 
 module.exports = {
   mode: "production",
-  entry: "./playground/app",
+  devtool: "source-map",
+  entry: [
+    "./src/index"
+  ],
   output: {
-    path: path.join(__dirname, "build"),
-    filename: "bundle.js",
-    publicPath: process.env.SHOW_NETLIFY_BADGE ? "": "/react-jsonschema-form/"
+    path: path.join(__dirname, "dist"),
   },
   plugins: [
     new MonacoWebpackPlugin({
       languages: ['json']
     }),
-    new MiniCssExtractPlugin({filename: "styles.css", allChunks: true}),
     new webpack.DefinePlugin({
       "process.env": {
-        NODE_ENV: JSON.stringify("production"),
-        SHOW_NETLIFY_BADGE: JSON.stringify(process.env.SHOW_NETLIFY_BADGE)
+        NODE_ENV: JSON.stringify("production")
       }
     })
   ],
-  resolve: {
-    extensions: [".js", ".jsx", ".css"]
-  },
   module: {
     rules: [
       {
@@ -37,22 +32,21 @@ module.exports = {
           path.join(__dirname, "src"),
           path.join(__dirname, "playground"),
           path.join(__dirname, "node_modules", "codemirror", "mode", "javascript"),
-        ],
+        ]
       },
       {
-        test: /\.css$/,
+        test: /\.s?css$/,
         use: [
-          {
-            loader: MiniCssExtractPlugin.loader,
-          },
+          "style-loader",
           "css-loader",
+          "sass-loader"
         ],
         include: [
-          path.join(__dirname, "css"),
+          path.join(__dirname, "src"),
           path.join(__dirname, "playground"),
           path.join(__dirname, "node_modules"),
         ],
-      }
+      },
     ]
   }
 };
