+++ packages/playground/webpack.config.prod.js
@@ -1,0 +1,60 @@
+var path = require("path");
+var webpack = require("webpack");
+const MonacoWebpackPlugin = require('monaco-editor-webpack-plugin');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+module.exports = {
+  mode: "production",
+  entry: "./playground/app",
+  output: {
+    path: path.join(__dirname, "build"),
+    filename: "bundle.js",
+    publicPath: process.env.SHOW_NETLIFY_BADGE ? "": "/react-jsonschema-form/"
+  },
+  plugins: [
+    new MonacoWebpackPlugin({
+      languages: ['json']
+    }),
+    new webpack.DefinePlugin({
+      "process.env": {
+        NODE_ENV: JSON.stringify("production"),
+        SHOW_NETLIFY_BADGE: JSON.stringify(process.env.SHOW_NETLIFY_BADGE)
+      }
+    }),
+    new HtmlWebpackPlugin({
+      filename: 'index.html',
+      template: 'playground/index.html'
+    })
+  ],
+  resolve: {
+    extensions: [".js", ".jsx", ".css"]
+  },
+  module: {
+    rules: [
+      {
+        test: /\.jsx?$/,
+        use: [
+          "babel-loader",
+        ],
+        include: [
+          path.join(__dirname, "src"),
+          path.join(__dirname, "playground"),
+          path.join(__dirname, "node_modules", "codemirror", "mode", "javascript"),
+        ],
+      },
+      {
+        test: /\.s?css$/,
+        use: [
+          "style-loader",
+          "css-loader",
+          "sass-loader"
+        ],
+        include: [
+          path.join(__dirname, "src"),
+          path.join(__dirname, "playground"),
+          path.join(__dirname, "node_modules"),
+        ],
+      }
+    ]
+  }
+};
